requirements_basket.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1702571698061/work
beautifulsoup4==4.11.1
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1696630167146/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
clip==0.2.0
colorama==0.4.6
cupy-cuda116==10.6.0
cycler==0.11.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dgl==1.1.3
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
faiss==1.5.3
fastrlock==0.8.2
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gdown==4.5.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1653252814274/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1637959344872/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912939357/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240553635/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1660929733321/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
joblib==1.2.0
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1710632238570/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1669810440410/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-client==5.3.4
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter_core==4.12.0
jupyterlab==1.2.21
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver==1.4.5
lightning-utilities==0.10.1
littleutils==0.2.2
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1654705952371/work
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1673904152039/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.5
munch==2.5.0
munkres==1.1.4
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
network==0.1
networkx==2.6.3
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1628216317882/work
numpy==1.21.6
ogb==1.3.6
outdated==0.2.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.3.5
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1694617248815/work
ply==3.11
prettytable==3.5.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1674535637125/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.1
psutil==5.9.5
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1663830492333/work
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole-base_1693604303222/work
QtPy @ file:///home/conda/feedstock_root/build_artifacts/qtpy_1698112029416/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.0.2
scipy==1.7.3
sconf==0.2.5
seaborn==0.12.2
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1682601222253/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.2.post1
tabulate==0.9.0
tensorboardX==2.5.1
termcolor==2.3.0
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1670253674810/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1645693042223/work
threadpoolctl==3.1.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
torch==1.12.1
torch-geometric==2.3.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torchaudio==0.12.1
torchmetrics==0.11.4
torchvision @ file:///croot/torchvision_1670313549871/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1656937818679/work
tqdm==4.64.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1699959196938/work
webencodings @ file:///home/conda/feedstock_root/build_artifacts/webencodings_1694681268211/work
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1647446886548/work
yacs==0.1.8
yarl==1.9.4
zipp==3.10.0
